fix(tests): separate mocks and tests
When Mocha runs on GitHub Actions, it tries to run the mock files to
look for tests... this triggers a syntax error. Instead, separate the
tests into their own directory, so that they can be treated in isolation
by Mocha.
